Add formation model, new simulation parameters
The new simulation parameters demonstrate the full behavior of the
ORCA-AFQ model. It includes three UAVs, one of which is the leader
of a formation.

The leader converges to a distant point, like in the previous
situation, but the others converge to a relative position above it.
The followers' paths conflict with the leader's, so ORCA acts upon
their desired velocities. Yet, after a relatively short time both
followers manage to get through and into their appropriate
positions within the formation. This demonstrates how ORCA does not
prevent convergence, as long as the convergence controller is
robust enough to work within any forced path.

The formation model is set to coincide with the leader's state, but
it can also be bound to any trackable reference point. This is not
directly supported, but is easily implemented by manually updating
the formation's position and velocity upon each simulation step.

Interestingly, the model is realistic enough to generate sensorial
imprecisions, which can be observed by as a slight instability in
how the quadrotors move. Still, ORCA is robust enough to make
collisions very hard, even under poor sensorial data, which can be
observed by how all agents in the formation tend to waver together.